requirement.txt
h5py
pandas
scipy
torchinfo
tqdm